% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_header.R
\docType{data}
\name{full_gt3x_header}
\alias{full_gt3x_header}
\title{Full Header for GT3X files}
\format{
A raw vector of values for the header
}
\usage{
full_gt3x_header
}
\description{
Full Header for GT3X files
}
\keyword{datasets}
