% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_header.R
\docType{data}
\name{gt3x_header}
\alias{gt3x_header}
\title{Header for GT3X files}
\format{
A raw vector of values for the header
}
\usage{
gt3x_header
}
\description{
Header for GT3X files
}
\keyword{datasets}
