% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_acc_csv.R
\name{write_actigraph_csv}
\alias{write_actigraph_csv}
\title{Write a CSV file for Actigraph}
\usage{
write_actigraph_csv(
  df,
  file = tempfile(fileext = ".csv"),
  sample_rate = 30L,
  max_g = c("8", "6")
)
}
\arguments{
\item{df}{A \code{data.frame} of \code{time}, and \code{X}, \code{Y}, \code{Z}}

\item{file}{output CSV file}

\item{sample_rate}{sampling frequency of the data}

\item{max_g}{max value (in g) of the device the data was from. If unsure,
use \code{"8"}}

\item{...}{additional arguments to pass to \code{\link[readr:write_delim]{readr::write_csv()}}}
}
\value{
The CSV file path output.
}
\description{
Write a CSV file for Actigraph
}
\examples{
if (requireNamespace("read.gt3x", quietly = TRUE)) {
  gt3xfile <- system.file(
    "extdata", "TAS1H30182785_2019-09-17.gt3x",
    package = "read.gt3x")
  df <- read.gt3x::read.gt3x(
    gt3xfile, imputeZeroes = FALSE, asDataFrame = TRUE,
    verbose = TRUE)
  sample_rate = attr(df, "sample_rate")
  max_g = attr(df, "acceleration_max")
  max_g = as.character(as.integer(max_g))
  out = write_actigraph_csv(df = df, sample_rate = sample_rate, max_g = max_g,
                            num_threads = 1)
}
}
