% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gt3x.R
\name{write_gt3x}
\alias{write_gt3x}
\title{Write a GT3X file}
\usage{
write_gt3x(
  df,
  file = tempfile(fileext = ".gt3x"),
  sample_rate = 30L,
  max_g = c("8", "6"),
  acceleration_scale = NULL
)
}
\arguments{
\item{df}{A \code{data.frame} of \code{time}, and \code{X}, \code{Y}, \code{Z}}

\item{file}{output GT3X file}

\item{sample_rate}{sampling frequency of the data}

\item{max_g}{max value (in g) of the device the data was from. If unsure,
use \code{"8"}}

\item{acceleration_scale}{The rescaling factor to scale the data to integers.
Normally, this is determined by the \code{max_g} (technically the serial
number/model of the device).  NOTE: this may result in a malformed GT3X file}
}
\value{
The GT3X file path output.
}
\description{
Write a GT3X file
}
\examples{
if (requireNamespace("read.gt3x", quietly = TRUE)) {
  gt3xfile <- system.file(
    "extdata", "TAS1H30182785_2019-09-17.gt3x",
    package = "read.gt3x")
  df <- read.gt3x::read.gt3x(
    gt3xfile, imputeZeroes = FALSE, asDataFrame = TRUE,
    verbose = TRUE)
  sample_rate = attr(df, "sample_rate")
  max_g = attr(df, "acceleration_max")
  max_g = as.character(as.integer(max_g))
  out = write_gt3x(df = df, sample_rate = sample_rate, max_g = max_g)
  read.gt3x::is_gt3x(out)
  read.gt3x::have_log_and_info(out)
  x <- read.gt3x::read.gt3x(out, imputeZeroes = FALSE, asDataFrame = TRUE,
                            verbose = TRUE)
  df_x = as.data.frame(x)
  df_df = as.data.frame(df)
  stopifnot(isTRUE(
    all.equal(
      df_x[, c("time", "X", "Y", "Z")],
      df_df[, c("time", "X", "Y", "Z")]
    )
  ))
  testthat::expect_warning(write_gt3x(df = df,
                                      sample_rate = sample_rate,
                                      max_g = "6"))
}
}
